requirements.txt
aliyun-python-sdk-core==2.11.4
aliyun-python-sdk-core-v3==2.11.4
asn1crypto==0.24.0
cache==1.0.3
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
construct==2.5.3
cryptography==2.3
Django==2.0
django-haystack==2.8.1
django-shortuuidfield==0.1.3
djangorestframework==3.9.2
future==0.17.1
idna==2.7
jieba==0.39
pefile==2018.8.8
Pillow==5.4.0
pycparser==2.18
PyMySQL==0.9.2
python-ptrace==0.9.3
pytz==2018.5
qiniu==7.2.2
requests==2.21.0
shortuuid==0.5.0
six==1.11.0
urllib3==1.24.1
Whoosh==2.7.4
xlrd==1.2.0
